Engine: Begin replacing glBegin/glEnd with vertex arrays. A step towards OpenGL ES support (and also faster).
git-svn-id: svn://svn.code.sf.net/p/opensludge/code@470 27a91dd4-ec73-4ff0-838b-64dfa05f697e

+++ source/Engine/graphics.h
@@ -26,7 +26,7 @@ struct shaders {
 
 // From Backdrop.cpp, but they're here anyway
 extern GLubyte * backdropTexture;
-extern double backdropTexW, backdropTexH;
+extern GLfloat backdropTexW, backdropTexH;
 
 
 extern unsigned int winWidth, winHeight;
